fix: #368 pull up convert method
